- a few minor bug fixes related to code warnings - code cleanup based on additional gcc warnings - enabled additional gcc compiler warnings

+++ source/glest_game/ai/path_finder.h
@@ -213,7 +213,7 @@ private:
 	TravelState aStar(Unit *unit, const Vec2i &finalPos, bool inBailout,
 			int frameIndex, int maxNodeCount=-1,uint32 *searched_node_count=NULL);
 	inline static Node *newNode(FactionState &faction, int maxNodeCount) {
-		if( faction.nodePoolCount < faction.nodePool.size() &&
+		if( faction.nodePoolCount < (int)faction.nodePool.size() &&
 			faction.nodePoolCount < maxNodeCount) {
 			Node *node= &(faction.nodePool[faction.nodePoolCount]);
 			node->clear();
